+++ dangminhtuan/odoo_training/views/sale_order_view.xml
@@ -15,6 +15,20 @@
             </field>
         </record>
 
+        <!-- Inherit the existing search view for sale order -->
+        <record id="sale_order_search_view_inherit" model="ir.ui.view">
+            <field name="name">sale.order.search.inherit</field>
+            <field name="model">sale.order</field>
+            <field name="inherit_id" ref="sale.view_sales_order_filter"/>
+            <field name="arch" type="xml">
+                <xpath expr="//search" position="inside">
+                    <filter string="Has tag"
+                            name="has_tag"
+                            domain="[('tag_ids', '!=', False)]"/>
+                </xpath>
+            </field>
+        </record>
+
         <!-- Define the action for the Order Form -->
         <record id="action_sale_order" model="ir.actions.act_window">
             <field name="name">Order Form</field>
